MonteCarloScattering

API Reference

Constants

MonteCarloScattering.constants.B_CMB0 — Constant

Equivalent B field to CMB energy density at a redshift of 0

MonteCarloScattering.constants.E₀_π⁰ — Constant

Neutral pion rest energy

MonteCarloScattering.constants.Mᵣₑₛ — Constant

Resonance mass

MonteCarloScattering.constants.T_CMB0 — Constant

Temperature of CMB at a redshift of 0

MonteCarloScattering.constants.Tₜₕ — Constant

Threshold kinetic energy, for pion production

MonteCarloScattering.constants.rad_loss_fac — Constant

Factor related to radiative losses

MonteCarloScattering.constants.Γᵣₑₛ — Constant

Resonance width

MonteCarloScattering.parameters.E_rel_pt — Constant

Cutoff for nonrelativistic vs relativistic equations, for particles

MonteCarloScattering.parameters.na_c — Constant

Max number elements in pcut array

MonteCarloScattering.parameters.na_cr — Constant

Maximum number of thermal particle crossings to hold before writing to scratch file

MonteCarloScattering.parameters.na_particles — Constant

Max number of particles at each pcut

MonteCarloScattering.parameters.na_photons — Constant

Maximum size of photon array

MonteCarloScattering.parameters.num_therm_bins — Constant

Number of bins in thermal distribution

MonteCarloScattering.parameters.psd_max — Constant

Maximum number of bins usable for phase space distribution and associated calculations. Applies to both momentum and angular dimensions

MonteCarloScattering.parameters.β_rel_fl — Constant

Cutoff for nonrelativistic vs relativistic equations, for fluid

Functions

MonteCarloScattering.F_stream! — Method

F_stream!(...)

TODO

MonteCarloScattering.adjacent_apply! — Method

adjacent_apply!(f, y, x)

In-place version of adjacent_apply.

MonteCarloScattering.adjacent_apply — Method

adjacent_apply(f, x)

Return f(x[i], x[i+1]) for each i in x's index, excluding the last i.

MonteCarloScattering.all_flux! — Method

all_flux!(...)

Tracks particle flux due to motion on the grid. Also finds number of new grid zone

Arguments

i_prt: Current particle number

i_grid_feb: free escape boundary index at grid

inj: if particle has been injected into acceleration process

aa: mass of particle, in units of proton mass

pb_pf: momentum of particle along magnetic field, in the plasma frame

p_perp_b_pf: momentum of particle perpendicular to magnetic field, in the plasma frame

ptot_pf: total momentum of particle, in the plasma frame

γₚ_pf: Lorentz factor of particle in the plasma frame

φ_rad:

weight:

uₓ_sk: bulk fluid velocity in the x-direction, in the shock frame

uz_sk: bulk fluid velocity in the z-direction, in the shock frame

utot: total bulk fluid velocity, in the shock frame

γᵤ_sf: Lorentz factor of fluid bulk motion in the shock frame

b_cosθ:

b_sinθ:

x_PT_cm:

x_PT_old:

Modifies

i_grid

i_grid_old

n_cr_count

num_crossings (array modified in-place)

pₓ_esc_upstream

energy_esc_upstream

pxx_flux (array modified in-place)

pxz_flux (array modified in-place)

energy_flux (array modified in-place)

spectra_sf (array modified in-place)

spectra_pf (array modified in-place)

psd (array modified in-place)

MonteCarloScattering.get_psd_bin_angle — Method

get_psd_bin_angle(...)

Given parallel component and total of a particle's momentum in the shock frame, determine which bin of psd particle will fall into.

Binning is done by the value of ptot_sk in code units. Angular binning is done with cos(θ) for large angles or θ for small angles.

The value of θ_fine marks the division between linear bin spacing (above θ_fine) and logarithmic spacing (from θ_fine down to θ_min). In both logarithmic regions the parameter bins_per_decade_*** determines the fineness of the bins.

Warning

binning in angle is actually done with the NEGATIVE of the particle's cosine. This lets the most finely spaced bins correspond to upstream-pointing particles rather than downstream-pointing ones. This also has the effect that angles are essentially measured from the -x axis rather than the +x axis.

Arguments

pₓ_sk: component of momentum parallel to shock normal (to B-field?), in code units

ptot_sk: total particle momentum, in code units

psd_bins_per_dec_θ

num_psd_θ_bins

psd_cos_fine

Δcos

psd_θ_min

Returns

Bin in angle into which particle falls

MonteCarloScattering.get_psd_bin_momentum — Method

get_psd_bin_momentum(...)

Given parallel component and total of a particle's momentum in the shock frame, determine which bin of psd particle will fall into.

Binning is done by the value of ptot_sk in code units. All binning is logarithmic.

Arguments

pₓ_sk: component of momentum parallel to shock normal (to B-field?), in code units

ptot_sk: total particle momentum, in code units

Returns

Bin in momentum into which particle falls

MonteCarloScattering.identify_corners — Method

identify_corners(...)

Given particular cell in PSD, identify the corners with lowest and highest total momenta and the lower/higher of the two remaining cos(θ) values.

Arguments

i: marker for location on the total momentum axis

j: marker for location on the cos(θ) axis

transform_corner_pt: values of total momentum for the corners of PSD, transformed into desired frame

transform_corner_cθ: values of cos(θ) for the corners of PSD, transformed into desired frame

ii_sk_pf: integer giving frame into which corners were transformed

Returns

pt_lo_pt: momentum of the corner with lowest total momentum

pt_lo_cθ: cos(θ) of the corner with lowest total momentum

pt_hi_pt: momentum of the corner with highest total momentum

pt_hi_cθ: cos(θ) of the corner with highest total momentum

cθ_lo_pt: momentum of the remaining corner with lower cos(θ)

cθ_lo_cθ: cos(θ) of the remaining corner with lower cos(θ)

cθ_hi_pt: momentum of the remaining corner with higher cos(θ)

cθ_hi_cθ: cos(θ) of the remaining corner with higher cos(θ)

pt_lo_tied: flag used if multiple cells are tied for lowest ptot

pt_hi_tied: flag used if multiple cells are tied for highest ptot

MonteCarloScattering.lorentz — Method

γ = lorentz(v)
γ = lorentz(β)

Get Lorentz factor γ from velocity v or β (in units of c).

MonteCarloScattering.main_loops — Method

main_loops(...)

Main computational loops Loops are set in the following sequence of nesting

loopitr: iiter Iteration number

loopion: iion Particle species number

looppcut: ipcut Particle splitting (pcut) number

looppt: iprt Individual particle number

MonteCarloScattering.new_pcut — Method

new_pcut(...)

Takes the **_saved arrays filled over the course of loop_pt and splits the saved particles to form the population of the next pcut

Arguments

TODO

n_pts_target

n_saved

grid_saved

tcut_saved

l_save

downstream_saved

inj_saved

weight_saved

ptot_pf_saved

pb_pf_saved

x_PT_cm_saved

xn_per_saved

prp_x_cm_saved

acctime_sec_saved

φ_rad_saved

n_pts_use

Returns

TODO

Modifies

weight_running: weight factor of each particle remaining after this pcut

MonteCarloScattering.particle_finish! — Method

particle_finish!(...)

Handles particles that leave the system during loop_helix for any reason.

Arguments

aa: particle atomic mass

pb_pf: component of ptot_pf parallel to magnetic field

p_perp_b_pf: component of ptot_pf perpendicular to magnetic field

γₚ_pf: Lorentz factor associated with ptot_pf

φ_rad: phase angle of gyration; looking upstream, counts clockwise from +z axis

uₓ_sk: bulk flow speed along x axis

uz_sk: bulk flow speed along z axis

utot: total bulk flow speed

γᵤ_sf: Lorentz factor associated with utot

b_cosθ: component of magnetic field along x axis

b_sinθ: component of magnetic field along z axis

i_reason: integer reason for why particle left system

i_iter

i_ion

weight: particle's weight

num_psd_mom_bins, num_psd_θ_bins

psd_bins_per_dec_mom, psd_bins_per_dec_θ

psd_mom_min, psd_θ_min

psd_cos_fine, Δcos,

mc

Modifies

pₓ_esc_feb

energy_esc_feb

esc_energy_eff

esc_num_eff

esc_flux

esc_psd_feb_downstream

esc_psd_feb_upstream

Returns

Nothing; modifies input argument arrays as needed.

MonteCarloScattering.print_input — Method

print_input(...)

Prints a whole mess of information to the screen and to mc_out

Arguments

n_pts_inj: target number of particles for injection distribution

n_pts_pcut: target number of particle for low-energy pcuts

n_pts_pcut_hi: target number of particles for hi-energy pcuts

n_ions: number of ion species run

num_psd_mom_bins: number of bins in momentum directions in psd

num_psd_θ_bins: number of bins in θ directions in psd

n_xspec: number of additional locations to track particle spectra

n_pcuts: number of pcuts

n_grid: number of grid zone BOUNDARIES, including x_grid_stop

r_RH: Rankine-Hugoniot compression ratio for this shock

r_comp: the compression ratio being used

u/β/γ ₀/₂: total fluid velocity[cm/s, /c] and Lorentz factor for far upstream and downstream regions

Returns

Nothing (it's all to the screen/file)

MonteCarloScattering.print_plot_vals — Method

print_plot_vals(...)

Prints the long list of floats at the end of each data set that are read in by the plotting program to display information about the run.

Arguments

iunit: unit number to which the line will be written

TODO

MonteCarloScattering.prob_return — Method

prob_return(...)

If the particle ends its movement downstream of the shock, perform a series of tests to determine whether it will be culled from the simulation.

Arguments

B_CMBz: effective magnetic field due to CMB at redshift of source; only used to pass to function retro_time when needed

x_PT_old: particle position before most recent move

aa: atomic mass number of ion species

gyro_denom: denominator of gyroradius fraction, q⋅B

helix_count: counter for number of times through main propagation loop for current particle

i_pcut: current pcut; needed when electrons are undergoing radiative losses

pcut_prev: momentum of previous pcut; needed when electrons are undergoing radiative losses

weight: current particle weight; passed to retro_time if called

Returns

i_return: flag for fate of particle; see top of loop_helix

lose_pt: true if particle hit zero energy due to radiative losses while in downstream region

Modifies

x_PT_cm: position of particle after recent motion & downstream adjustment

prp_x_cm: location of probability-of-return plane

ptot_pf: total plasma frame momentum of particle

γ_pf: Lorentz factor associated with ptot_pf

gyro_denom: denominator of gyroradius fraction, q⋅B

pb_pf/p_perp_b_pf: components of ptot_pf parallel/perpendicular to B field

acctime_sec: total accumulated acceleration time

φ_rad: phase angle of particle's gyration

tcut_curr: current tcut for particle tracking; passed to retro_time if called

MonteCarloScattering.q_esc_calcs — Method

q_esc_calcs(...)

Use the Rankine-Hugoniot relations to calculate the escaping momentum & energy flux.

MonteCarloScattering.q_esc_calcs_nonrelativistic — Method

q_esc_calcs_nonrelativistic(...)

Solution comes from Ellison (1985) 1985JGR....90...29E (Eqs 8-10). Note assumption of zero escaping momentum flux, which is good to within a couple percent for strong nonrelativistic shocks. #TODO: check how much of a difference this assumption makes

MonteCarloScattering.q_esc_calcs_relativistic — Method

q_esc_calcs_relativistic(...)

Solution comes from Ellison+ (1990) [1991ApJ...378..214E]. Uses relativistic Rankine-Hugoniot relations. See that paper for details of equations and associated quantities. Briefly,

\[\begin{align} g_0 n_0 b_0 &= g_2 n_2 b_2 \tag{RH1} \\ g_0^2 w_0 b_0^2 + P_0 &= g_2^2 w_2 b_2^2 + P_2 + Q_{px} \tag{RH2} \\ g_0^2 w_0 b_0 c &= g_2^2 w_2 b_2 c + Q_{en} \tag{RH3} \end{align}\]

where

$w = E_0 + E_k + P$ is the enthalpy as total energy density + pressure

$E_0 = nmc^2$ is the rest mass energy density

$E_k = (γ-1) nmc^2$ is the kinetic energy density, with $γ = \sqrt{1 + (p/mc)^2}$

$P = \frac{1}{3} npv$ is the pressure

For closure, it is assumed that the two escaping fluxes are related by

\[Q_{en} = \sqrt{(1+β₀)/2} ⋅ Q_{px} ⋅ c,\]

i.e. the geometric mean of the arithmetic mean of u₀ and c. This allows the solution to smoothly join with the non-relativistic version. Use only fluid component of fluxes, not fluid+EM, for now.

MonteCarloScattering.radiation_loss — Method

radiation_loss(B², p, Δt)

TODO

MonteCarloScattering.retro_time — Method

retro_time(...)

Explicitly tracks particles downstream of the probability of return plane. Does so by the "retrodictive approach" explained in Jones (1978) [1978ApJ...222.1097J] and also by Ellison, Jones, & Reynolds (1990) [1990ApJ...360..702E].

Since we know the particle will return to the PRP, we run time "backwards" by placing the particle in a bulk flow opposite to its motion. Given a presumably infinite downstream region of propagation, the particle is guaranteed to return to the PRP over long enough time scales, and we assume that the "true" history of the particle (its PRP1→downstream→PRP2 path) resembles on average its backwards history (PRP1←downstream←PRP2).

Arguments

B_CMBz: effective magnetic field due to CMB at redshift of source

aa: atomic mass number of ion species

zz: charge of ion species

prp_x_cm: starting (and ending) location for retro_time motion

weight: current particle weight

Returns

lose_pt: whether particle is lost due to energy losses despite "returning" from a probabilistic standpoint

φ_rad: phase angle of particle upon return

Modifies

ptot_pf: total plasma frame momentum of particle

pb_pf/p_perp_b_pf: components of ptot_pf parallel/perpendicular to B field

γ_pf: Lorentz factor associated with ptot_pf

gyro_denom: denominator of gyroradius fraction, zz*B

acctime_sec: total accumulated acceleration time

tcut_curr: current tcut for particle tracking

MonteCarloScattering.scattering — Method

scattering(...)

This is a combination of two subroutines from the old code: prob_scat and scattering. Randomly moves the particle's momentum vector along the surface of the unit sphere.

Arguments

aa: particle atomic mass

gyro_denom: q*B, the denominator of gyroradius formula

ptot_pf: total plasma frame momentum

γₚ_pf: Lorentz factor associated with ptot_pf

xn_per: number of time steps a gyroperiod is divided into

Returns

gyro_period_sec

Modifies

pb_pf: component of ptot_pf parallel to magnetic field

p_perp_b_pf: component of ptot_pf perpendicular to magnetic field

φ_rad: phase angle of gyration

MonteCarloScattering.tcut_print — Method

tcut_print(...)

If tcuts were tracked during the run, print out the particle counts and spectra at each tcut

Arguments

FIXME

weights_fileunit

spectra_fileunit

n_tcuts: Number of times to use in particle tracking

tcuts: Array to hold cutoff times for particle tracking

n_ions: Number of ion species

MonteCarloScattering.tcut_track! — Method

tcut_track!(...)

If directed in data_input, track particles' momenta – and whether they are still interacting with the shock – at various times since acceleration began.

Modifies

weight_coupled

spectra_coupled

Arguments

tcut_curr: current tcut for tracking

weight: particle weight

ptot_pf: plasma-frame total momentum

i_ion

num_psd_mom_bins

Returns

Nothing. All adjustments made to input arrays

MonteCarloScattering.thermo_calcs — Method

thermo_calcs(...)

Reads in the PSD – as well as the thermal arrays & scratch file – to calculate the pressure (which may be anisotropic) everywhere on the grid.

Arguments

num_crossings: array containing number of thermal particle crossings recorded at each grid zone

n_cr_count: total number (up to na_cr) of thermal particle crossings; signals whether scratch file was used

therm_grid: array of grid zone number for thermal crossings

therm_pₓ_sk: array of x-momentum for thermal crossings

therm_ptot_sk: array of total momentum for thermal crossings

therm_weight: array of flux-adjusted particle weights for thermal particle crossings

nc_unit: unit number for the scratch file holding crossing data

psd: 3-D phase space distribution holding shock-frame ptot and cos(θ) for all recorded CR grid crossings

zone_pop: (Lorentz-invariant) number of particles in each grid zone

Returns

P_psd_par: component of pressure parallel to magnetic field

P_psd_perp: component of pressure perpendicular to magnetic field

energy_density_psd: local kinetic energy density of fluid in every grid zone

MonteCarloScattering.β — Method

β(γ)

Get velocity β (in units of c) from Lorentz factor γ.

MonteCarloScattering.cosmo_calc.get_redshift — Method

get_redshift(d_CM)

Calculator to get redshift from comoving distance. Subroutine adapted from Hogg (1999). For more detail see Hogg (1999) [DOI 10.48550/arXiv.astro-ph/9905116]

Arguments

d_CM: Comoving distance in Mpc

Returns

z: Redshift

MonteCarloScattering.initializers.F_update! — Method

F_update!(...)

Update the flux arrays as if the particles had actually crossed them.

Arguments

pxx_flux: Modified in-place.

pxz_flux: Modified in-place.

energy_flux: Modified in-place.

m_ion

n₀_ion

T₀_ion

relativistic

i_stop

u₀

γ₀

γ_sf_grid

uₓ_sk_grid

MonteCarloScattering.initializers.alfven_speed — Method

alfven_speed(Val(:classical), ρ, B)

Alfvén wave group velocity.

MonteCarloScattering.initializers.alfven_speed — Method

alfven_speed(Val(:relativistic), ρ, B, P, Γ)

Calculate the Alfvén speed for a relativistic plasma with pressure P, density ρ, Lorentz factor Γ, and ambient magnetic field B.

It uses the Equation (46) given by Gedalin (1993), where we assume an equation of state $e = ρc^2 + P/(Γ-1)$.

\[\begin{align*} v_A &= c \sqrt{\frac{u_B}{ε + p + u_B}} \\ &= \frac{c}{\sqrt{1 + \frac{w}{u_B}}} \end{align*}\]

where $u_B = B^2/4π$ is the magnetic field energy density, $w = \frac{Γ}{Γ-1} P + ρ c^2$ is the enthalpy density, $ε$ is ??, $p$ is ??, $e$ is the (total internal) energy density, $ρ c^2$ is the rest energy density, and $P/(Γ-1)$ is the thermal component (internal kinetic energy).

MonteCarloScattering.initializers.calc_MB_area — Method

calc_MB_area(p_range, E_range)

Area under the curve for a Maxwell-Boltzmann distribution, up to arbitrary normalization.

Arguments

p_range

E_range

Returns

area_tot

MonteCarloScattering.initializers.calc_MB_area_single_bin — Method

calc_MB_area_single_bin(p1, p2, E1, E2)

Area under the curve for a single bin of Maxwell-Boltzmann distribution, up to arbitrary normalization. Uses trapezoid rule for integration.

Arguments

p1, p2: Momentum bounds of bin

E1, E2: Energy bounds of bin (assumes rel/nonrel calculations are already done), in units of kT

Returns

The area occupied on the bin under a Maxwell-Boltzmann distribution of temperature T (the temperature is implicit in E1 and E2).

MonteCarloScattering.initializers.calc_downstream — Method

calc_downstream(...)

Uses the Rankine-Hugoniot jump conditions to calculate the downstream conditions for a test particle shock. Big difference between this subroutine and calc_rRH is that we already know what the downstream speed is, courtesy of r_comp in the input.

Shock must be parallel (not oblique).

Arguments

B₀: far upstream magnetic field strength [Gauss]

r_comp: compression ratio of shock

β₀: upstream bulk fluid speed, over c

Returns

β: (total) bulk fluid speed downstream

γ: Lorentz factor associated with β₂

B: downstream magnetic field strength [Gauss]

θ_B: angle[deg] between downstream magnetic field and shock normal

θᵤ: angle[deg] between downstream fluid velocity and shock normal

MonteCarloScattering.initializers.calc_rRH — Method

calc_rRH(...)

Uses the Rankine-Hugoniot jump conditions to calculate the compression ratio for a shock assuming test-particle conditions. In other words, (1) sharp shock, (2) negligible/no DSA, and (3) no escaping flux. Additionally assumes that the inflowing plasma has non-relativistic thermal speeds to make upstream adiabatic index exactly 5/3.

Arguments

β₀, γ₀: shock speed and Lorentz factor

n_ions: number of different ion species

m_ion: array of particle species' mass

n₀_ion: array of far upstream densities for particle species

T₀_ion: array of particle species' far upstream temperatures

Shock must be parallel (not oblique)

Returns

r_RH: Rankine-Hugoniot compression ratio

Γ₂_RH: ratio of specific heats (adiabatic index) for downstream region, assuming r_comp = r_RH

MonteCarloScattering.initializers.calc_rRH_nonrelativistic — Method

calc_rRH_nonrelativistic(P₀, ρ₀, β₀)

Nonrelativistic, parallel.

Solution comes from Ellison (1985) [1985JGR....90...29E]. Uses far upstream Mach number to calculate r_RH.

MonteCarloScattering.initializers.calc_rRH_relativistic — Method

calc_rRH_relativistic(species, ρ₀, P₀, β₀, n₀_ion)

Relativistic, parallel.

Solution comes from Ellison & Reynolds (1990) 1991ApJ...378..214E. Uses relativistic Rankine-Hugoniot relations. See that paper for details of equations and associated quantities. Briefly,

\[\begin{align*} γ_1 n_1 b_1 &= γ_2 n_2 b_2 \tag{R-H1} \\ γ_1^2 w_1 b_1^2 + P_1 &= γ_2^2 w_2 b_2^2 + P_2 \tag{R-H2} \\ γ_1^2 w_1 b_1 &= γ_2^2 w_2 b_2 \tag{R-H3} \end{align*}\]

where

$w = e_0 + e_K + P$ is the enthalpy as total energy density + pressure

$e_0 = n m c^2$ is the rest mass energy density

$e_K = n m c^2 (γ - 1)$ is the kinetic energy density, with $γ = \sqrt{1 + (p/mc)^2}$

$P = \frac{1}{3} n p v$ is the pressure

Assumes that downstream particle distributions are δ-functions. Solves for p₂ using Newton's method, then works backwards to r_RH.

MonteCarloScattering.initializers.create_inj_dist_momentum_range — Method

Create range of momenta over which the Maxwell-Boltzmann particle distribution is set.

Arguments

T: distribution temperature[K]

m: mass for this particle species

Returns

p_range: Array of momenta

MonteCarloScattering.initializers.init_pop — Method

init_pop(...)

Initializes the particle populations that will propagate through the shock structure. Handles fast push and associated flux-tracking & changes to the population

Arguments

do_fast_push

inp_distr

i_ion

m

rng

T₀_ion

energy_inj

inj_weight

n_pts_inj

n₀_ion

x_grid_start

rg₀

η_mfp

x_fast_stop_rg

β₀, γ₀, u₀

n_ions

m_ion

n_grid

x_grid_rg

uₓ_sk_grid

γ_sf_grid

ptot_inj

weight_inj

n_pts_MB

Returns

n_pts_use

i_grid_in

weight_in

ptot_pf_in

pb_pf_in

x_PT_cm_in

pxx_flux

pxz_flux

energy_flux

MonteCarloScattering.initializers.mach_alfven_func — Method

mach_alfven_func(u, P, ρ, Γ, B, relativistic)

Return Alfvénic mach number (ratio of speed to Alfvén wave group velocity)

MonteCarloScattering.initializers.mach_sonic_func — Method

mach_sonic_func(u, P, ρ, Γ, relativistic)

Return sonic mach number (ratio of speed to local speed of sound)

MonteCarloScattering.initializers.set_custom_εB! — Method

set_custom_εB!(...)

Arguments

εB_grid

grid_axis

n_ions

species

B₀

F_px_upstream

F_energy_upstream

uₓ_sk_grid

x_grid_rg

comp_fac

γ₀, β₀, u₀

MonteCarloScattering.initializers.set_downstream_photon_shells! — Method

set_downstream_photon_shells!(...)

TODO

MonteCarloScattering.initializers.set_inj_dist — Method

set_inj_dist(inj_weight, n_pts_inj, inp_distr, T_or_E, m, n₀)

Sets the injected particle distributions for all species. Initially, particles are placed in a Maxwell-Boltzmann (thermal) distribution based on supplied temperature and particle mass. This is corrected at the end if a δ-function distribution was requested (not worried about the wasted computation because this subroutine runs only rarely).

Arguments

inj_weight: whether each particle or each bin has equal weights (true for equal weight particles, false for equal weight bins)

n_pts_inj: target # of particles for distribution

inp_distr: thermal, δ-function, or some other distribution

T_or_E: if using thermal distribution, this is temperature[K]; if δ function, it's injection energy[keV]

m: mass for this particle species

n₀: far upstream number density for this species

Returns

ptot_out: array holding plasma frame total momenta for all particles in the distribution

weight_out: array holding particle weights

n_pts_use: number of particles in the distribution; will almost surely be different from n_pts_inj if using thermal distribution

CHECKTHIS: that output distribution matches M-B, just to make sure I haven't made a typo

MonteCarloScattering.initializers.set_inj_dist_bin_equal_weight! — Method

set_inj_dist_bin_equal_weight!(...)

Arguments

ptot_out: Modified in-place.

weight_out: Modified in-place.

p_range

E_range

area_tot

Δp

n₀

n_per_bin

Returns

ptot_out

weight_out

MonteCarloScattering.initializers.set_photon_shells — Method

set_photon_shells(...)

If photon calculation is desired, photons will be collected into upstream and downstream shells for easier viewing. This subroutine sets the endpoints of the shells, as well as their midpoints.

Because the particle spectrum changes most rapidly near the shock, zones should be small near the shock and get larger as you move further away.

First, divide the domain between the shock and the FEB (either upstream or downstream into n sections, where n is the respective number of shells to use. HOWEVER, do this by exponent, ranging from -1 to log10(|x_FEB|).

Keep track of the boundaries of each shells, since we will need those for calculating the total number of particles emitting when we get to that point in photon production. The midpoints are less useful, since photons are calculated on a zone-by-zone basis rather than just at select points in the shock profile. Keep them anyway, since the memory overhead is low.

MonteCarloScattering.initializers.set_psd_angle_bins — Method

set_psd_angle_bins(...)

Sets the BOUNDARIES of the bins of the phase space distribution. The bins are numbered from 0 to num_psd_θ_bins, each boundary denotes the lower edge of that # bin; the indices thus run from 0 to num_psd_θ_bins + 1.

Warning

the number stored in psdθbounds increases at first (increasing θ), then decreases because increasing the angle means decreasing the cosine.

Linear spacing of cosine values for angles between psdθfine and π. Below psdθfine spacing is logarithmic in θ for some number of decades down to psdθmin. Values less than the minimum are equivalent to 0.0.

Arguments

psd_bins_per_dec_θ: number of bins per decade to use in logarithmically-spaced regions of PSD

psd_lin_cos_bins: number of bins to divide range [-1,cos(psdθfine)] into

psd_cos_fine: cutoff between lin/cos and log/θ spacing of PSD bins

psd_θ_min: minimum angle for PSD

Returns

num_psd_θ_bins: total number of bins along given dimension, not counting bin 0

Δcos: size of each linear cosine bin

psd_θ_bounds: boundaries between bins, and upper edge of final bin

MonteCarloScattering.initializers.set_psd_mom_bins — Method

set_psd_mom_bins(...)

Sets the BOUNDARIES of the bins of the phase space distribution. The bins are numbered from 0 to num_psd_mom_bins, each boundary denotes the lower edge of that # bin; the indices thus run from 0 to num_psd_mom_bins + 1.

Logarithmic spacing over all decades from E_min to E_max is used. Values less than the minimum are equivalent to 0.0.

Arguments

psd_mom_min: minimum momentum to use in PSD

psd_mom_max: maximum momentum to use in PSD

psd_bins_per_dec_mom: # of bins per decade to use in logarithmically-spaced regions of PSD

Returns

num_psd_mom_bins: total number of bins along given dimension, not counting bin 0

psd_mom_bounds: boundaries between bins, and upper edge of final bin

MonteCarloScattering.initializers.set_upstream_photon_shells! — Method

set_upstream_photon_shells!(...)

TODO

MonteCarloScattering.initializers.setup_grid — Method

setup_grid(...)

Arguments

x_grid_start_rg

x_grid_stop_rg

use_prp

feb_downstream

rg₀

Returns

x_grid_rg

x_grid_start

x_grid_stop

MonteCarloScattering.initializers.setup_profile — Method

setup_profile(...)

Sets the initial values of the shock profile

Arguments

u₀, β₀, γ₀

B₀

θ_B₀

r_comp

bturb_comp_frac

bfield_amp

use_custom_εB,

n_ions

species

F_px_upstream

F_energy_upstream

grid_axis

x_grid_cm

x_grid_rg

Returns

uₓ_sk_grid: bulk fluid velocity along x axis (i.e., perpendicular to shock face) in shock frame

uz_sk_grid: bulk fluid velocity along z axis (i.e., parallel to shock face) in shock frame

utot_grid: total bulk fluid velocity in shock frame

γ_sf_grid: bulk flow Lorentz factor in shock frame

β_ef_grid: relative x-axis speed between plasma and explosion frames

γ_ef_grid: Lorentz factor associated with βefgrid

btot_grid: total magnetic field

θ_grid: angle of magnetic field[radians] relative to shock normal (i.e., to x axis)

εB_grid: user-defined function for fraction of energy density in magnetic field. Sets value of btot_grid

B₂: field strength in downstream region. Initially set in calcdownstream, it may be reset here depending on values of bturbcompfrac & bfieldamp

MonteCarloScattering.initializers.sound_speed — Function

sound_speed(regime, P, ρ, Γ)

Return the speed of sound in a plasma.

Arguments

regime: Must be one of Val(:classical) or Val(:relativistic)

P: plasma pressure

ρ: plasma mass density

Γ: plasma adiabatic index

MonteCarloScattering.initializers.upstream_energy_flux — Function

upstream_energy_flux(regime=Val(:classical), u₀, β₀, ρ₀, P₀, Γ, B_z)
upstream_energy_flux(regime=Val(:relativistic), u₀, β₀, γ₀, e₀, ρ₀, P₀, B_z)

Arguments

regime: Whether to use Newtonian or relativistic calculations

u₀, β₀, γ₀

e₀

ρ₀

P₀

B₀, B_x, B_z

Γ: Adiabatic index

MonteCarloScattering.initializers.upstream_fluxes — Method

upstream_fluxes(...)

Calculates the far upstream fluxes for the shock.

Two different cases considered:

Non-relativistic oblique shock. Uses equations of Ellison+ (1996) [1996ApJ...473.1029E]

Relativistic shock, any obliquity. Uses equations of Double+ (2004) [2004ApJ...600..485D]

Only oblique equations used because they reduce trivially to parallel cases when θ_B₀ = 0.

HOWEVER, assumes that z-component of far upstream velocity is 0 in all cases; in practice oblique shocks would induce some z-velocity in the shock profile even though particles initially arrive with no bulk z component. Also assumes isotropic initial pressure, so no off-diagonal components in pressure tensor.

Shock must be parallel (not oblique). Set oblique to false.

Arguments

n₀_ion

T₀_ion

m_ion

B₀

θ_B₀

u₀, β₀, γ₀

No inputs; pulls everything from module 'controls'

Returns

F_px_upstream: far upstream momentum flux, x component

F_pz_upstream: far upstream momentum flux, z component

F_energy_upstream: far upstream energy flux

MonteCarloScattering.initializers.upstream_machs — Method

upstream_machs(β₀, species, B₀)

Calculates the sonic and Alfvén mach numbers for the shock.

For speed of sound, uses Equation (13) of Fujimura & Kennel (1979) [1979A%26A....79..299F]

For Alfvén wave speed, uses Equation (46) of Gedalin (1993) [1993PhRvE..47.4354G]

Arguments

β₀: Far upstream plasma speed

species: Vector of species in the plasma (of type Species)

B₀: Far upstream magnetic field

Returns

mach_sonic: sonic Mach number of upstream flow

mach_alfven: Alfvénic Mach number of upstream flow

MonteCarloScattering.initializers.upstream_momentum_flux — Function

upstream_momentum_flux(regime=Val(:classical), u₀, β₀, ρ₀, P₀, B_x, B_z, Γ)
upstream_momentum_flux(regime=Val(:relativistic), β₀, γ₀, e₀, P₀, B₀, B_x, B_z)

Arguments

regime: Whether to use Newtonian or relativistic calculations

u₀, β₀, γ₀

e₀

ρ₀

P₀

B₀, B_x, B_z

Γ: Adiabatic index

MonteCarloScattering.particle_counter.cos_centers — Method

Get the cosine of the centers of the θ bins of the phase space distribution

MonteCarloScattering.particle_counter.get_dNdp_2D — Method

get_dNdp_2D(...)

Generates d²N/(dp dcos) (or something like it) for electrons in the explosion frame.

First, use the scratch file and the phase space distribution to generate d²N/(dp dcos), the 2-D version of dN/dp (with extra information about the angular distribution). Combine the results from thermal and CR particles into a single d²N/(dp dcos) to save memory. After d²N/(dp dcos) is found in the shock frame, transform it to the plasma and ISM frames by re-binning the center of each shock frame bin. Leave the computationally difficult problem of overlap for a later date. Finally, condense each d²N/(dp dcos) into dN/dp for comparison against the results of the original dN/dp subroutines.

Warning

d²N/(dp dcos) isn't technically accurate. To get number, just multiply by dp and sum over the cos(θ) column.

Note

Right now, this is only used to calculate the ISM-frame distribution of electrons; all other frame/particle pairs can be done with the 1-D dN/dp subroutines.

Arguments

nc_unit: unit number for the scratch file holding crossing data

zone_pop: (Lorentz-invariant) number of particles in each grid zone

Returns

d²N_dpdcos_ef: explosion-frame array holding dN/dp spread out across angular dimension

MonteCarloScattering.particle_counter.get_dNdp_cr — Method

get_dNdp_cr(...)

Calculate dN/dp (NOT normalized) in plasma frame for particles that have been injected into acceleration process.

Arguments

FIXME

None, but does use PSD information from module psd_vars

Returns

dNdp_cr: 3-D array, containing 1-D array for each grid zone, of dN/dp for particles injected into acceleration process, for three different reference frames

MonteCarloScattering.particle_counter.get_dNdp_therm — Method

get_dNdp_therm(...)

Calculate dN/dp (NOT normalized) for particles that haven't been injected into acceleration process.

First create arrays to hold crossing information from this ion species, and fill them with data from the scratch file. Next, loop over grid locations, performing three main tasks:

Transform shock frame values into plasma frame

Seek out maximum and minimum momenta in plasma frame

Bin crossings according to plasma frame total momentum

Also perform the above process for the ISM frame, and store shock frame values for comparison against earlier results.

Arguments

num_hist_bins: number of bins to use in histograms

nc_unit: unit number of scratch file for crossing data

Returns

dNdp_therm: 3-D array, containing 1-D array for each grid zone of dN/dp for the thermal particles

dNdp_therm_pvals: array of momentum bin boundaries for each row of dNdp_therm; each row handled separately to maximize resolution of what may be an extremely narrow peak at radically different energy from the upstream population

MonteCarloScattering.particle_counter.get_normalized_dNdp — Method

get_normalized_dNdp(...)

Computes the actual number of particles in each bin of dN/dp (which is divided by dp, remember). Computes total area under curve, normalizes it against number of particles upstream using plasma-frame density and volume, and then uses fractional area of each bin to determine number of particles in it. Handles non-injected (i.e. thermal) particles differently than injected ones, due to very small p.f. spread in momenta.

The array that would be dNdp_cr_pvals is already set, as psd_mom_bounds

Arguments

nc_unit: unit number for the scratch file holding crossing data

Returns

dNdp_therm: 3-D array, containing 1-D array for each grid zone of dN/dp for the thermal particles

dNdp_therm_pvals: array of momentum bin boundaries for each row of dNdp_therm; each row handled separately to maximize resolution of what may be an extremely narrow peak at radically different energy from the upstream population

dNdp_cr: 3-D array, containing 1-D array for each grid zone of dN/dp for the population of accelerated particles

zone_pop: (Lorentz-invariant) number of particles in each grid zone

MonteCarloScattering.particle_counter.rebin_dNdp_therm — Method

rebin_dNdp_therm(num_hist_bins, dNdp_therm, dNdp_therm_pvals, num_psd_mom_bins, psd_mom_bounds)

This subroutine takes the distribution of thermal particles and re-bins it in the bins used for the cosmic rays. Note that the thermal distribution is provided (and re-binned separately) in all three frames: shock, plasma, and ISM.

Arguments

num_hist_bins: number of bins used for the thermal distribution

dNdp_therm: distribution of thermal particles in all three frames

dNdp_therm_pvals: arrays holding bin boundary values for the thermal distributions; as with the distributions themselves, they change based on grid location

num_psd_mom_bins: number of bins used for the cosmic ray distribution

psd_mom_bounds: boundaries of the bins for the cosmic ray distribution Unlike the thermal boundaries, these are constant everywhere on the grid.

Returns

Re-binned histogram of the thermal distribution

MonteCarloScattering.particle_counter.set_grid_volumes! — Method

set_grid_volumes!(zone_pop, zone_vol, x_grid_cm, i_shock, γ₀, jet_sph_frac)

Find the volume and population of each grid zone in x_grid_cm.

Arguments

x_grid_cm

i_shock

γ₀

jet_sph_frac

uₓ_sk_grid

γ_sf_grid

Modifies

zone_pop

zone_vol

MonteCarloScattering.smoothers.smooth_grid_par — Method

smooth_grid_par(...)

Uses tracked fluxes of particles, and the Rankine-Hugoniot jump conditions, to determine the smoothed shock profile for the next iteration of the code. Only valid in parallel case, which makes subroutine, inputs and equations simpler than would be required for oblique case.

Arguments

FIXME

i_iter: current iteration number

n_grid: number of grid zones

x_grid_rg: locations[rg₀] of grid zone boundaries

uz_sk_grid: z-component of bulk fluid velocity, in shock frame. This is a parallel shock, so uz_sk_grid = 0 identically; just used for output

θ_grid: angle[rad] between mean magnetic field and shock normal. This is a parallel shock, so θ_grid = 0 identically; just used for output

pxx_flux: momentum flux of particles across grid zone boundaries

energy_flux: energy flux of particles across grid zone boundaries

Γ₂: downstream adiabatic index

F_px_upstream: far upstream momentum flux, calculated in upstream_fluxes

F_energy_upstream: far upstream energy flux, calculated in upstream_fluxes

Returns

utot_grid: total bulk flow speed of fluid in shock frame (is a pure output because uₓ_sk_grid is used in the calculations instead)

γ_ef_grid: Lorentz factor of flow relative to far upstream plasma (i.e. in the explosion frame)

β_ef_grid: bulk flow speed associated with γ_ef_grid

Modifies

uₓ_sk_grid: x-component of bulk fluid velocity, in shock frame

γ_sf_grid: Lorentz factor associated with utot_grid, but since this is a parallel shock it's the Lorentz factor associated with uₓ_sk_grid

btot_grid: magnetic field strength[G] in each grid zone, including any compression of turbulence or additional amplification

MonteCarloScattering.smoothers.smooth_profile! — Method

smooth_profile!(y_prof, n_grid)

Takes an input velocity profile and performs two tasks: enforces monotonicity by smoothing out dips/bumps, and averages nearby points to smooth sharp edges

Arguments

y_prof: array holding velocity profile (modified in-place)

n_grid: number of grid zones in position and velocity arrays

MonteCarloScattering.transformers.get_transform_dN — Method

get_transform_dN(...)

Calculate dN(p) (a 1-D array) for the passed slice of PSD in the specified inertial frame.

Arguments

psd: the (2-D) slice of the larger phase space distribution

m: integer specifying frame into which we're transforming

transform_corner_**: array holding transformed corner values, both ptot and cos(θ)

γ: conversion factor from number density to flux

i_approx: degree of approximation to use in computing the dNdp_out

Returns

dN_out: dN(p) for the given slice of PSD once transformed into the specified frame

MonteCarloScattering.transformers.track_pitch_angles — Method

track_pitch_angles(...)

TODO

MonteCarloScattering.transformers.transform_p_PS — Method

transform_p_PS(...)

Takes particle's Plasma frame momentum components and transforms them to the shock frame. Transformation formulas are correct for all obliquities.

Notes from Glen Double (31 Oct 2002): Method: define flow velocity and particle momentum in component vectors with respect to the plasma xyz frame. Find p_para (with respect to u) component of particle momentum using scalar product of p and u. Then p_perp = p - p_para (using vectors). Next, make relativistic transformation on ppara while `pperpremains constant. Finally, recreateprel = prelpara + pperp, and findp_rel` components in the original xyz frame by taking scalar products along the xyz axes.

Arguments

aa: particle atomic mass

pb_pf: component of ptot_pf parallel to magnetic field

p_perp_b_pf: component of ptot_pf perpendicular to magnetic field

γ_pf: Lorentz factor associated with ptot_pf

φ_rad: phase angle of gyration; looking upstream, counts clockwise from +z axis

uₓ_sk: bulk flow speed along x axis

uz_sk: bulk flow speed along z axis

utot: total bulk flow speed

γ_sf: Lorentz factor associated with utot

b_cosθ: component of magnetic field along x axis

b_sinθ: component of magnetic field along z axis

Returns

ptot_sk: total shock frame momentum in new grid zone

p_sk: components of shock frame momentum

γ_sk: Lorentz factor associated with ptot_sk

MonteCarloScattering.transformers.transform_p_PSP — Method

transform_p_PSP(...)

Takes particle's old plasma frame momentum components and transforms them twice: from old Plasma frame to new Shock frame, then from new Shock frame to new Plasma frame. Only ever called if particle scattered between zones with different bulk flow velocities. Transformation formulas are correct for all obliquities.

Notes from Glen Double (31 Oct 2002): Method: define flow velocity and particle momentum in component vectors with respect to the plasma xyz frame. Find p_para (with respect to u) component of particle momentum using scalar product of p and u. Then p_perp = p - p_para (using vectors). Next, make relativistic transformation on p_para while p_perp remains constant. Finally, recreate p_rel = p_rel_para + p_perp, and find p_rel components in the original xyz frame by taking scalar products along the xyz axes.

Arguments

aa: particle atomic mass

uₓ_sk/old: current and old bulk flow speed along x axis

uz_sk/old: current and old bulk flow speed along z axis

utot/old: current and old total bulk flow speed

γ_sf/old: Lorentz factor associated with utot/old

b_cosθ/old: current and old component of magnetic field along x axis

b_sinθ/old: current and old component of magnetic field along z axis

Returns

ptot_pf: total plasma frame momentum in new grid zone

ptot_sk: total shock frame momentum in new grid zone

p_sk: xyz components of ptot_sk

pb_sk: component of ptot_sk parallel to magnetic field

p_perp_b_sk: component of ptot_sk perpendicular to magnetic field

γₚ_sk: Lorentz factor associated with ptot_sk

γ_pf: Lorentz factor in plasma frame, associated with ptot_pf

Modifies

pb_pf: component of ptot_pf parallel to magnetic field

p_perp_b_pf: component of ptot_pf perpendicular to magnetic field

φ_rad: phase angle of gyration; looking upstream, counts clockwise from +z axis

MonteCarloScattering.transformers.transform_psd_corners — Method

transform_psd_corners(...)

Given a relative Lorentz factor between two frames, transform the corners of the PSD into the new frame. Outputs total momentum as log of cgs values.

TODO: remove conversion to/from log space here and in get_dNdp_CR

Arguments

γ: relative Lorentz factor between shock and resultant frame

E₀: Rest energy of particle

psd_lin_cos_bins

num_psd_θ_bins

psd_θ_bounds

psd_mom_bounds

Returns

transform_corner_pt: total momenta at the corners

transform_corner_ct: cos(θ) (NOT θ!!!) values at the corners

MonteCarloScattering.transformers.triangular_distribution! — Method

triangular_distribution!(...)

TODO

MonteCarloScattering.transformers.uniform_cell_distribution! — Method

uniform_cell_distribution!(...)

TODO

Bibliography

[1]

D. C. Ellison. Shock acceleration of diffuse ions at the earth's bow shock Acceleration efficiency and A/Z enhancement. Journal of Geophysical Research 90, 29–38 (1985-01-01). ADS Bibcode: 1985JGR....90...29E.

[2]

D. C. Ellison and S. P. Reynolds. A Determination of Relativistic Shock Jump Conditions Using Monte Carlo Techniques. The Astrophysical Journal 378, 214 (1991-09-01). ADS Bibcode: 1991ApJ...378..214E.